requirements.txt
asgiref==3.7.2
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.3.0
Django==4.2.9
django-import-export==3.3.6
django-mssql-backend==2.8.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
gunicorn==23.0.0
iniconfig==2.0.0
MarkupPy==1.14
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyodbc==5.0.1
pytest==7.4.4
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.8.0
xlrd==2.0.1
xlwt==1.3.0
